refactor: use constants barrel file (index.js) across all imports
- All imports now go through src/constants/index.js instead of
  directly to validation.js or ui.js
- constants/index.js already existed with export * from both files
  but was unused — now it serves its intended purpose
- 8 files updated: getAuthorNames, useCourseForm, useRegistrationForm,
  SearchBar, Courses, Header, CourseInfo, Input
- Shorter import paths, single entry point for constants module
